+++ pytorch_optimizer/optimizer/alig.py
@@ -8,6 +8,16 @@ from pytorch_optimizer.base.types import CLOSURE, DEFAULTS, LOSS, PARAMETERS
 from pytorch_optimizer.optimizer.utils import get_global_gradient_norm
 
 
+@torch.no_grad()
+def l2_projection(parameters: PARAMETERS, max_norm: float = 1e2):
+    r"""Get l2 normalized parameter."""
+    global_norm = torch.sqrt(sum(p.norm().pow(2) for p in parameters))
+    if global_norm > max_norm:
+        ratio = max_norm / global_norm
+        for param in parameters:
+            param.mul_(ratio)
+
+
 class AliG(BaseOptimizer):
     r"""Adaptive Learning Rates for Interpolation with Gradients.
