+++ pytorch_optimizer/optimizer/utils.py
@@ -321,16 +321,6 @@ def enable_running_stats(model):
 
 
 @torch.no_grad()
-def l2_projection(parameters: PARAMETERS, max_norm: float = 1e2):
-    r"""Get l2 normalized parameter."""
-    global_norm = torch.sqrt(sum(p.norm().pow(2) for p in parameters))
-    if global_norm > max_norm:
-        ratio = max_norm / global_norm
-        for param in parameters:
-            param.mul_(ratio)
-
-
-@torch.no_grad()
 def get_global_gradient_norm(param_groups: List[Dict]) -> torch.Tensor:
     r"""Get global gradient norm."""
     global_grad_norm = torch.zeros(1, dtype=torch.float32, device=param_groups[0]['params'][0].device)
